Fix screen migration of Scratch Pad
The bug also caused a crash when trying to change display
scaling on Wayland.

When a ScratchPad is created its parent, the brush editor, doesn't
yet have a native surface. Hence it was connected to the main
window instead, which caused the scratchpad track not the
current screen of the brush editor, but the current screen of
the main window.

The patch reuses the code that tracks the surface of for
KisRootSurfaceInfoProxy. This code was extracted into a separate
class KisRootSurfaceTrackerBase, which is now used in both places.
It tracks the whole hierarchy of scatch pad's parents and connects
to the actual toplevel widget dynamically.