[android] Enter fullscreen early to avoid flicker
Instead of waiting for the main window to do it, Krita now enters
fullscreen mode when the application starts up. This avoids the main
window spawning in at a size where it doesn't fully fill the screen,
only to then a moment later get resized in an ugly flicker fit.