Fix color picker preview activating incorrectly
There were a few weirdnesses in the color picker handling that could
lead to incorrect results.

If you hit Ctrl and then clicked before the delayed activation timer was
up, the color picker would start picking before the preview was
summoned, leading to it setting the foreground color to the picked color
and comparing to the wrong thing as a result. The same thing happened
pretty much every time when using the dedicated color picker tool.

It also didn't properly summon the comparison when just clicking, only
doing so when the cursor was moved slightly. The same also happened when
pressing and holding.

Both of these issues are now fixed by triggering the preview and marking
a sample present when the sampling action is started.