Don't show an error when cancelling an export
If the file type prompts you before actually saving (for example, PNG)
it's still possible to cancel the export process. However the previous
logic considers this a failure, and will show a pretty unhelpful dialog
that there was an "unknown failure" while exporting your document.

Instead, lets treat it like a cancelled export which doesn't display a
dialog. startExportInBackground unfortunately doesn't provide enough
information to the callee since a boolean can only state if it was
successful or not, I chose to just bubble up the KisImageExportErrorCode
instead.

BUG: 501128