Add a floating message for wrong use of Knife Tool
Before this commit, if the user tried to use a Knife Tool
on a paint layer or any other node, it would still show the normal
decorations, it would just not do any actions.
(The options would also suggest to create a vector layer first).
This commit adds a floating message, consistent with what
Default Tool does in the same situation.