Implement KoShapeBulkActionInterface
This interface allows shapes to postpone computation
intensive operations till the end of the user action.
The bulk-action mode also handles updates of the
dependent shapes.

The mode is currently supposed to be used in the text
shape to handle the updates coming from the contour
shapes.

See TestDependentShapes for the reference
implementation of the bulk mode handling.