Merge rects together in GridIterationTools
Before this commit, every rect was processed and therefore copied
at the time it was provided, which didn't allow any kind of optimization.
This commit saves them into a list and then processed them all at once
after feeding them into KisRegion to be merged together.
It needs to first check if it's okay to do so by ensuring
that all the polygons are convex (therefore not folded over themselves
or one another).