Refactor public API for "enter text group" functionality
1) Fix lifetime issues with the "group text shape" pointer in
KisCanvas2. Now the pointer is wrapped into a special
ShapeLifetimeWrapper that leaves the group as soon as
the shape is destroyed.

TODO: we need to make this wrapper a bit more generic, we
have the same code in KoSvgTextShape::Private

2) Rename public API to make it a little be more generic and
undestandable. We will be able to reuse it later for the group
shapes as well.

Before:
virtual KoSvgTextShape *textShapeManagerEnabled() const;
virtual void setTextShapeManagerEnabled(KoSvgTextShape  *source);

After:
virtual KoShape *currentShapeManagerOwnerShape() const;
virtual void setCurrentShapeManagerOwnerShape(KoShape *source);

They are basically the same functions, but are generic and can
be used for groups later.