Add cutOutSubgridFromBounds to GridIterationTools
Before this commit, there was no way to get the areas
around the cut out subgrid.
This commit adds this function.

This function allows the user of the Grid Iter Tools to
process the area around the subgrid differently from
the specific subgrid. It returns maximum four rectangles
(in pixel coordinates) around the subgrid, one from each side.
It skips empty rectangles.