Use SpatialContainer in LiquifyTransformWorker
Before this commit, LiquifyTransformWorker would use normal
lists of points, and to move them, it would go through the whole list.
This commit adds spatial containers to the worker, so instead of
going through the whole list, it only needs to find those points
in the semi-quadtree of the container, making it way faster.