PythonPluginManager: Fix crash on Python 3.14
I'm not an expert in Python reference counting, but we called Py_DECREF
*twice* (once in Python::functionCall, and then later in the callee)
which Python 3.14 really doesn't like. For some reason this creates a
confusing, random crash on start-up as we begin importing modules. This
crash is very easy to reproduce on Fedora Linux 43 (see
https://bugzilla.redhat.com/show_bug.cgi?id=2408113)

I only tested this fixed against Python 3.14 (which is way more modern
than the officially supported version, I know) this fix should be fine
for older versions - even if we don't crash there. Plus when Krita does
update to 3.14, then it won't crash :-)