[android] Make shortcut input work properly
On Android, releasing a modifier key will have the key release event
still carry the modifier that was release, causing the key sequence
input widget to get stuck thinking the user is holding the modifier for
all eternity. This commit handles that case and clears the modifiers
explicitly when their corresponding key is released.