Handle IME in on-canvas text input properly
It now actually shows the on-screen keyboard when the input is
interacted with, doesn't eternally get stuck in a "block IME query"
state, updates the IME positions when they happen instead of one
interaction later, updates the IME positions as the canvas is moved
around and doesn't update the IME positions when the text tool isn't
what's receiving inputs.

This now uses an RAII locker object to bock and unblock the IME queries,
since one of the bugs that caused these problems was due to the code
setting a flag and then forgetting to unset it in an early return.