Add segments and index conversion to VectorPath
Before this commit, walking through a shape was still
somewhat complicated due to the segmentAt() function
giving two points with some unrelated data.
It was also easy to mix up indexes of QPainterPath
and indexes of segments in VectorPath.

This commit adds a class Segment to allow easier
access of information for a specific segment,
and converting indexes between a QPainterPath and segments.