Add isInsideShape to KisAlgebra2D
Before this commit, there was no way to determine is a point is inside a
QPainterPath or VectorPath shape.
This commit adds it together with all the supporting functions
for finding intersections in bezier curves (using BezierUtils).
The function is not optimized, because it doesn't seem to be needed,
but in case it needs to be optimized, do this:
- use just one horizontal line but properly check all the intersections
on the ends of the edges
- keep in mind that intersections on bezier curves can be very
  inaccurate (see the circle test)