Cancel removing gutter when going over a whole shape
Before this commit, if the gutter was crossing exactly two segments
on the same shape, the code would try to use it as if it was a gutter,
even if it was on a convex shape, going from outside to the outside.
This commit checks whether both end points of the crossing mouse line
are within the shape to ensure that it is the gutter that we're
removing.