Upgraded-KisToolGradient-Init-Capture-Lambda
`KisToolGradient` was allocating redundant memory on the stack which was rendered obsolete by the C++14 init-capture update for Lambda captures.

Prior to C++14, it was acceptable to copy data onto the stack from a class for the sole purpose of forwarding them into the Lambda capture at the cost of redundancy by copying data onto the stack.

The biggest benefit of init-capture is the readability and maintainability, especially when jumping around with IDEs, but it also creates a slightly smaller and faster runtime stack. It's also considered "safer", because init-captures align with C++ RAII.

According to the C++ reference documentation:

"if the capture is by-copy, the introduced non-static data member of the closure object is another way to refer to that variable;"

See "Lambda capture" section https://en.cppreference.com/w/cpp/language/lambda.html

Ticket: #None Keywords: TODO

## Test Plan

* Ran Krita and manually stress tested the gradient tool.
* objdump'ed the `.so` and validated the disassembly.
  * `objdump -dSlC /home/appimage/appimage-workspace/krita-build/bin/kritadefaulttools.so | grep endPrimary`

## Formalities Checklist

- [x] I confirmed this builds.
- [x] I confirmed Krita ran and the relevant functions work.
- [x] I tested the relevant unit tests and can confirm they are not broken. (If not possible, don't hesitate to ask for help!)
- [x] I made sure my commits build individually and have good descriptions as per [KDE guidelines](https://community.kde.org/Policies/Commit_Policy).
- [x] I made sure my code conforms to the standards set in the HACKING file.
- [x] I can confirm the code is licensed and attributed appropriately, and that unattributed code is mine, as per [KDE Licensing Policy](https://community.kde.org/Policies/Licensing_Policy).
- [ ] Does the patch add a user-visible feature? If yes, is there a documentation MR ready for it at [Krita Documentation Repository](https://invent.kde.org/documentation/docs-krita-org)?

**_Reminder: the reviewer is responsible for merging the patch, this is to ensure at the least two people can build the patch. In case a patch breaks the build, both the author and the reviewer should be contacted to fix the build._** **_If this is not possible, the commits shall be reverted, and a notification with the reasoning and any relevant logs shall be sent to the mailing list, _****_kimageshop@kde.org_****_._**