Switch to coarse layer thumbnails when generation is slow
When a layer has vast areas of empty pixels, it may be slow
to calculate its exactBounds(). In this case, if thumbnail generation
takes more than 40ms, Krita will automatically switch the layer
into "coarse thumbnail generation mode" and use extent() instead.

The status bar will also show a warning symbol and list the "slow
layers" in a tooltip.

A proper solution to this problem will come later in this MR:
https://invent.kde.org/graphics/krita/-/merge_requests/2578

BUG:510627